many. xah-beginning-of-line-or-block and other. see below.
new xah-beginning-of-line-or-block and xah-end-of-line-or-block.

xah-beginning-of-line-or-block now move cursor to begining of first visible char. (i.e. after the indentation)

xah-beginning-of-line-or-block cursor result is always beginning (of line or indentation).
xah-end-of-line-or-block cursor result is always end of line.

xah-upcase-sentence fix. reverted previous (possibly uncommited) change

new xah-insert-form-feed. Now also insert empty line around it.

xah-insert-form-feed now has key 【space e l】

renamed xah-display-page-break-as-line to xah-display-form-feed-as-line

balance-windows now has key 【space 5】