removed key qwerty 【space w】 for universal-argument, because
the command universal-argument is a none-repeat call command. So, we give it 【space w】

But the C-u is repeatable key, the default. So, not a good choice.

But universal-argument is often useful when in insert mode to insert many chars. This means, user has to switch to insert mode, which means 【space w】 key is not available.

so, this create a situation where sometimes you need to do 【space w】 and other times you need to do 【control+u】.

solution is just stick with the default C-you.