SPC h is , basically,help?-- calling it 'help'
I notice that SPC h is , basically,help.  But the function call is 'xah-fly-h-keymap', which is what comes up in the minibuffer. I've changed this to 'help'. There are a number of other similar changes that can be contemplated. My guess is that indeed, these function have names that are just temporary placeholders. The idea would be to use some organizing principles for superficial discoverability. Yet I don't think I can discern the patters of 'xah-fly-w-keymap, and 'xah-fly-n-keymap', and so forth. xah-fly-w-keymap and then a 'p', for example, is eval-expression. So, SPC-w-p is eval-expression. While SPC-w-w is 'delete frame'. I don't see what these have to do with each other, or what they have to do with 'w'. I am not, I suppose, mincing words, here, but is this supposed to make some sort of sense? Does it seem important to hot-key 'delete frame' at all? Maybe, but less, do I need to be able to hot-key 'eval expression'. Also, goto-line is SPC-t-f. It so happens I was just looking at avi line, and there is also an avi word, avi-pop-mark, avi-timer, avi-goto-url. So I think of grouping them together, mapping them to a leader prefix. Somewhere. I don't mean to be peremptory, here, or at least, not brusquely imperious. I'm just sharing the thinking rattling around in my head..