there is a subtle issue local vs. global state
if I switch from 'insert' mode to 'command' mode, and vice versa, things seem to be working. But if I enter the minibuffer interactively, things go into 'insert' mode'. Probably shouldn't be a 'global' state', of being in 'insert' or 'command' mode, and shouldn't switch just because I go into the minibuffer interatively (which is always in 'insert' mode, but shouldn't affect mode of other buffers). This isn't quite perfect. The specific issue that I notice is if I am in 'command' mode, and then co  into the minibuffer interatively. This switches me to 'insert' mode. Then I leave the minibuffer. I seem to revert to 'command' mode, which is correct, but the modeline isn't updated to 'command' mode flag. This is subtle, the description of the issue is confusing here -- it seems like the keybindings of 'insert' mode and 'command' mode are always correct, but the mode is not correct about what mode I am in. Reproduce by going into 'command' mode, and then going into the minibuffer interactively, by hitting 'M-x', and then just going 'C-g' to exit the minibuffer again. This puts me in command mode, but the modeline incorrectly shows me as being in 'insert' mode.