status -- working
debugged re: issues mentioned in last commit. I titled that commit like this: 'there is a subtle issue local vs. global state', and said 'if I enter the minibuffer interactively, things go into 'insert' mode'. Probably shouldn't be a 'global' state'..' etc. What I have done, is incorporated used of 'make-local-variable' command, to manipulate the command and insert keymaps to be associated with local buffers. This means that when you interactively go into the minibuffer, you can cleanly return to the buffer you were in, and it's still in the 'command' or 'insert' state that it was in.