updated to turn off those I/C hints if 'mode off'
My previous commit said 'status: working'.

A large claim. I add a further commit here, because I see this:  'the modeline still has I/C hints' if you disable the xah-fly-keys-mode. 

Therefore, I have included just a line to turn off those I/C hints. More generally, see the whole state of the past discussion about 'why it's not turning off', and when/if the mode should turn off automatically. I have made changes here. The whole issue of 'fixing the mode on/off...' has a few aspects. In other words, how to 'turn xfk on/off properly'. I am currently happy but I've made multiple changes related to this one issue, and it may well be that something in the new behavior is a surprise. If we just step back and list what we want, we want to turn the mode on and off cleanly. No error messages, no incorrect minibuffer messages. I have to admit I've made multiple changes here.. one may well debate some decisions. One change that I experimented with, but will only mention, is that a globalized minor mode is a global minor mode that is created from an existing (non-global) minor mode. This idea caught my attention. Vaguely, my impression is that if something is not a global minor mode, then it needs to be enabled on a per-buffer basis. If you *genuinely* want it to be global, everywhere, you can easily define a global minor mode yourself. It's simple, I mean, there is this idea of a global minor mode. And if this is a good approach, then this is because you can toggle it on and off globally. Just something that I came across, but it wasn't being used and I haven't added it.