made a globalized minor mode
I reference my notes on my previous commit -- about trying to write a global minor mode. I want to understand exactly how, for example, define-minor-mode is a macro. And there is a difference between a minor-mode, and a globalized minor mode. Previously, xah-fly-keys was using the 'regular' (local) minor mode.  I'd like to use the globalized minor mode, and have made that change. I think it does work perfectly, but still, it didn't necessarily have to be changed. Is there a 'right way'?  A kind of complicated question (at least for me). I don't know exactly what one will want, if there are changes in behavior because I have configured a globalized minor mode that I use everywhere (it has a high precedence). I am sure something will come up, but in the meantime, I have concentrated on testing how you can interactively turn xah-fly-keys on and off, using 'xah-fly-keys-mode-on', and 'xah-fly-keys-mode-off'. 

It seems to work perfectly for me. 

I have actually changed the way the minibuffer is handled, too. It does not have an 'insert' or 'command' mode. Previous behavior of xah-fly-keys was to put the minibuffer in 'insert' mode, but I'm just not enabling 'xah-fly-keys' for the minibuffer at all. This helps with cleanly turning xah-fly-keys on and off cleanly, and updating the minibuffer cleanly as to what mode we are in, 'command' or 'insert'. The minibuffer itself is not in either mode. This is an 'obvious' change, I think, in retrospect. Logical. Also, it proves that I can proceed to include logic to decide whether xah-fly-keys should be enabled for different major modes.