make sure to clean up on xah-fly-command-mode-init subsequent calls
`xah-fly-command-mode-init` uses a feature of the `set-transient-map`
that explicitly disables a clean up on a pre-command-hook
invocation (`keep-pred` parameter):

```
(setq xah-fly--deactivate-command-mode-func
        (set-transient-map xah-fly-command-map (lambda () t)))
```

In our case it is `(lambda () t)`.

From the code of the `set-transient-map`:

```
However, if the optional argument KEEP-PRED is t, MAP stays
active if a key from MAP is used.  KEEP-PRED can also be a
function of no arguments: it is called from `pre-command-hook' and
if it returns non-nil, then MAP stays active.
```

Although `xah-fly-command-map` map stays active until insert mode is
started (as it should be since we don't want quite a command mode
prematurely) and XFK takes responsibility of invoking the cleanup
function (`xah-fly--deactivate-command-mode-func`) when insert mode is
activated, `xah-fly-command-mode-init` could be called multiple times.

Subsequent activation of the command mode overwrites
`xah-fly--deactivate-command-mode-func`, meaning that we loose a
reference to a clean up func that is supposed to remove the
`clear-transient-map` items from the `pre-command-hook`.

This commit ensures that a subsequent command mode activation clears
up a transient map set by already active command mode.