changed repeat key to backspace, was -, for repeatable commands xah-reformat-to-sentence-lines, xah-cycle-hyphen-lowline-space
new var xah-repeat-key if not defined.

key space return is now back to execute-extended-command, not execute-extended-command-for-buffer (the buffer one does not allow you to call other commands (e.g. vc-diff, useful in lots programing language modes), make it much less useful, because you need to know which command is or is not for current major mode etc. it seems actually slower too)

now C-y is undo-redo of emacs 28.

removed alias execute-extended-command-for-buffer for execute-extended-command for emacs less that 28. (since we no longer use execute-extended-command-for-buffer. the code was incorrect anyway. The version comparison code was incorrect.)