major internal change. * New. var xah-fly-layouts holds a list of layouts. so, layouts can be programatically listed.
* function xah-fly-keys-set-layout major change. no longer support M-x customize. Reason: supporting it makes the code much more complex.

* var xah-fly-key-current-layout changed accordingly.

* renamed xah-fly--current-layout-kmap to xah-fly--key-convert-table

* here is some detail. the layout was variables, eg:

xah--dvorak-to-azerty-kmap
xah--dvorak-to-azerty-be-kmap
xah--dvorak-to-beopy-kmap
etc.

now these variables are string keys of xah-fly-layouts, with prefx xah--dvorak-to- dropped and suffix -kmap dropped, so the names are just azerty, azerty-be, etc. value remain the same.

* starting to phase out using symbol for layout names.
Symbol pollutes symbol table. and string is more intuitive, universal.
e.g.
(xah-fly-keys-set-layout 'qwerty)
should be
(xah-fly-keys-set-layout "qwerty")