Experimental. now ESC key by default isbound to xah-fly-command-mode-activate. Reason:
Advantage:

* on Kinesis 360 http://xahlee.info/kbd/kinesis_360.html and Glove80 Keyboard http://xahlee.info/kbd/glove80_keyboard.html the traditional capslock position is Esc. This makes it convenient for these keyboards.

* also, it's a convention of vi/vim/neovim to have esc activate command mode. This can save some learning time.

* save the trouble to rebind capslock to home key, for those want efficient key.

The disadvantage of bind esc:

* On text terminal, in the case that meta key is not setup, you lose the ability to press meta key bound commands. (but this is rare on modern terminals.)

* Esc, on traditonal keyboard, is a bit far away, and not optimal. So, lazy people may develope a habit to turn left wrist to press it. This is a Repetitive Strain Injury danger.

* On some people's xah-fly-keys customization, Esc is remapped to C-g (cancel) by key-translation-map. for those who did this, it'll still work.

Note:

on ergonomic keyboards and layouts, there is an emerging standard to not have a capslock key on the traditional position to the left of qwerty A. (e.g. google chromebook, and many batman keyboards)

but there is no emerging standard what that key should be.
colemak has backspace (but on ergonomic keyboards with thumb keys, that's not a optimal choice)
some has esc. some has ctrl. (not a good choice unless on laptop)