xah-fly-command-mode-activate now bound to <escape> too. (works in gui emacs only.) reason:
advantage of making esc key do activate command mode:

- familiar to most people
- one less key. (if you use home key, it's an extra key on your physical keyboard but otherwise the Home key function isn't much used these days.)

disadvantage:

- esc by default is at hard to press spot. induces Repetitive Strain Injury.
- by emacs quirk, <escape> means esc key in gui mode only. while ESC means esc key in both gui and terminal, but binding ESC, will break arrow keys and f1 f2 etc keys, due to how terminal works (terminal keys are actually esc sequences).

in the end, i decided to bind it. you should esc key position on your keyboard capslock position, or, when pressing esc on the left corner, don't twist your wrist, but move whole hand to it. and, when in terminal, no choice but to press other keys to activate command mode, such as M-SPC.