experimental. 【space m】 now is a leader key. was 「dired-jump」. here are the new keys. (all dvorak label) ("m m" . dired-jump) ("m e" . delete-other-windows) ("m u" . split-window-below) ("m w" . universal-argument)
reason: wanted the split window commands to have a leader key. part of the reason is that describe-function does not list the key of a command that is available in command mode. e.g. describe-function on delete-other-windows wont show the key 3 while in command mode. (this is because since xah-fly-keys use transient mode keys)

the other reason is that, the number row keys require some reach. now you have alternative of key sequence. not necessarily better, but having alternative keys are usually good. (pressing the key 3 every few minutes for about 7 years got me a specific RSI)